+++ sample/src/main/java/com/davidecirillo/multichoicesample/sampleToolbar/SampleToolbarActivity.java
@@ -4,6 +4,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.IntDef;
 import android.support.annotation.Nullable;
+import android.support.annotation.VisibleForTesting;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -21,6 +22,7 @@ import com.davidecirillo.multichoicesample.ResultActivity;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.util.ArrayList;
+import java.util.Random;
 
 import butterknife.BindView;
 import butterknife.OnClick;
@@ -30,9 +32,9 @@ public class SampleToolbarActivity extends BaseActivity {
     private static final int DEFAULT_QUANTITY_MODE = QuantityMode.STRING;
 
     @IntDef({
-            QuantityMode.NONE,
-            QuantityMode.STRING,
-            QuantityMode.PLURALS,
+        QuantityMode.NONE,
+        QuantityMode.STRING,
+        QuantityMode.PLURALS,
     })
     @Retention(RetentionPolicy.SOURCE)
     public @interface QuantityMode {
@@ -62,6 +64,18 @@ public class SampleToolbarActivity extends BaseActivity {
     @Override
     protected int setActivityIdentifier() {
         return R.layout.activity_sample_toolbar;
+    }
+
+    @OnClick(R.id.notify_data_changed)
+    void notifyDataChanged() {
+
+        stringList.clear();
+        Random random = new Random();
+        for (int i = 0; i < random.nextInt(30); i++) {
+            stringList.add("New item " + i);
+        }
+
+        mMySampleToolbarAdapter.notifyAdapterDataSetChanged();
     }
 
     @OnClick(R.id.result)
@@ -98,13 +112,13 @@ public class SampleToolbarActivity extends BaseActivity {
         stringList = getSampleList();
 
         MultiChoiceToolbar.Builder builder = new MultiChoiceToolbar.Builder(SampleToolbarActivity.this, toolbar)
-                .setMultiChoiceColours(R.color.colorPrimaryMulti, R.color.colorPrimaryDarkMulti)
-                .setDefaultIcon(R.drawable.ic_arrow_back_white_24dp, new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        onBackPressed();
-                    }
-                });
+            .setMultiChoiceColours(R.color.colorPrimaryMulti, R.color.colorPrimaryDarkMulti)
+            .setDefaultIcon(R.drawable.ic_arrow_back_white_24dp, new View.OnClickListener() {
+                @Override
+                public void onClick(View view) {
+                    onBackPressed();
+                }
+            });
 
         switch (quantityMode) {
             case QuantityMode.NONE:
@@ -161,15 +175,15 @@ public class SampleToolbarActivity extends BaseActivity {
                     boolean select = mMySampleToolbarAdapter.select(2);
                     if (!select) {
                         Toast.makeText(this, "Item not selected because not in multi choice mode or single click mode, select something first.",
-                                Toast.LENGTH_LONG).show();
+                            Toast.LENGTH_LONG).show();
                     }
                     return true;
 
                 case R.id.single_click_mode:
                     mMySampleToolbarAdapter.setSingleClickMode(!mMySampleToolbarAdapter.isInSingleClickMode());
                     Toast.makeText(getApplicationContext(),
-                            "Always Single Click Mode [" + mMySampleToolbarAdapter.isInSingleClickMode() + "]",
-                            Toast.LENGTH_SHORT).show();
+                        "Always Single Click Mode [" + mMySampleToolbarAdapter.isInSingleClickMode() + "]",
+                        Toast.LENGTH_SHORT).show();
 
                     return true;
 
@@ -199,10 +213,20 @@ public class SampleToolbarActivity extends BaseActivity {
         return quantityMode;
     }
 
+    @VisibleForTesting
+    public MySampleToolbarAdapter getMySampleToolbarAdapter() {
+        return mMySampleToolbarAdapter;
+    }
+
     public void setQuantityMode(@QuantityMode int quantityMode) {
         mMySampleToolbarAdapter.deselectAll();
 
         this.quantityMode = quantityMode;
         setUpMultiChoiceRecyclerView();
+    }
+
+    public void setStringList(ArrayList<String> list) {
+        stringList.clear();
+        stringList.addAll(list);
     }
 }